+++ DawnLib.Dusk/src/API/Definitions/EntityReplacement/Unlockables/DuskUnlockableReplacementDefinition.cs
@@ -1,0 +1,47 @@
+using Unity.Netcode;
+using UnityEngine;
+
+namespace Dusk;
+
+public abstract class DuskUnlockableReplacementDefinition : DuskEntityReplacementDefinition<DuskUnlockable>
+{
+}
+
+public abstract class DuskUnlockableReplacementDefinition<T> : DuskUnlockableReplacementDefinition where T : DuskUnlockable
+{
+    protected abstract void Apply(T dawnUnlockable);
+    public override void Apply(DuskUnlockable dawnUnlockable)
+    {
+        Apply((T)dawnUnlockable);
+        dawnUnlockable.SetUnlockableReplacement(this);
+        foreach (Hierarchy hierarchyReplacement in Replacements)
+        {
+            hierarchyReplacement.Apply(dawnUnlockable.transform);
+        }
+
+        foreach (GameObjectWithPath gameObjectAddon in GameObjectAddons)
+        {
+            if (string.IsNullOrEmpty(gameObjectAddon.PathToGameObject))
+            {
+                continue;
+            }
+
+            GameObject? gameObject = dawnUnlockable.transform.Find(gameObjectAddon.PathToGameObject)?.gameObject;
+            if (gameObject != null)
+            {
+                if (gameObjectAddon.GameObjectToCreate.TryGetComponent(out NetworkObject networkObject) && !NetworkManager.Singleton.IsServer)
+                    continue;
+
+                GameObject addOn = GameObject.Instantiate(gameObjectAddon.GameObjectToCreate, gameObject.transform);
+                addOn.transform.localPosition = gameObjectAddon.PositionOffset;
+                addOn.transform.rotation = Quaternion.Euler(gameObjectAddon.RotationOffset);
+
+                if (networkObject == null)
+                    continue;
+
+                networkObject.AutoObjectParentSync = false;
+                networkObject.Spawn();
+            }
+        }
+    }
+}