+++ Thunderstore/CHANGELOG.md
@@ -4,7 +4,7 @@
   - Also allows for the rotations of items to be retained in comparison to vanilla resetting rotations.
   - Also allows you to get rid of and add mods that add items without your save's items getting scrambled.
   - Added config to opt out of it in case of issues or dislike I guess.
-
+- Added UnlockableReplacementDefinition for replacing the model etc of unlockables
 ## v0.3.4
 
 - Fixed a small issue with custom moon terminal loading.